Revert "chore(suite-native): move SettingsSection to settings"
This reverts commit cedad4160969ffd8230bc92e9281221280c5eb39.
