+++ suite-native/module-settings/src/components/FeaturesSettings.tsx
@@ -4,17 +4,18 @@ import { useNavigation } from '@react-navigation/core';
 import { useAtomValue } from 'jotai';
 
 import {
-    RootStackParamList,
-    RootStackRoutes,
     SettingsStackRoutes,
+    RootStackRoutes,
     StackNavigationProps,
+    RootStackParamList,
 } from '@suite-native/navigation';
 import { FeatureFlag, useFeatureFlag } from '@suite-native/feature-flags';
 import { Translation } from '@suite-native/intl';
 import { selectHasDeviceDiscovery } from '@suite-common/wallet-core';
-import { SettingsSection, SettingsSectionItem } from '@suite-native/settings';
 
 import { useSettingsNavigateTo } from '../navigation/useSettingsNavigateTo';
+import { SettingsSection } from './SettingsSection';
+import { SettingsSectionItem } from './SettingsSectionItem';
 import { isDevButtonVisibleAtom } from './ProductionDebug';
 
 export const FeaturesSettings = () => {
