+++ suite-native/module-settings/src/components/PreferencesSettings.tsx
@@ -1,8 +1,9 @@
 import { Translation } from '@suite-native/intl';
 import { SettingsStackRoutes } from '@suite-native/navigation';
-import { SettingsSection, SettingsSectionItem } from '@suite-native/settings';
 
 import { useSettingsNavigateTo } from '../navigation/useSettingsNavigateTo';
+import { SettingsSection } from './SettingsSection';
+import { SettingsSectionItem } from './SettingsSectionItem';
 
 export const PreferencesSettings = () => {
     const navigateTo = useSettingsNavigateTo();
