+++ suite-native/module-settings/src/components/SettingsSection.tsx
