+++ suite-native/module-settings/src/components/SettingsSectionItemIcon.tsx
