+++ suite-native/settings/src/index.ts
@@ -1,3 +1,1 @@
 export * from './settingsSlice';
-export * from './components/SettingsSection';
-export * from './components/SettingsSectionItem';
